Moved tests to source directory.
