+++ botto/mocks/__init__.py
@@ -1,0 +1,1 @@
+from .mocks import MockRequest, MockObject
